Added STL exposing bug
The use of std::vector reveals bug in the freelist implementation.

Even without this bug, I think the STLAllocator is still broken. How it
is written, it shares the main allocator which causes infinite loops.
The STLAllocator needs to be written to use it's own source heap (which
is probably a singleton).