Create an heap for sqlite-enabled page table.
These changes add a sqlite-compatible memory allocator powered be a
seperate internal memory heap. This is a drop in replacement for
sqlite's memory allocator.

This gives us the ability to use a dedicated region of memory for
maintaining a database for our library. We'll start by using it to
manage the page table for application memory.