Fix delimiter bugs
The lines of a request don't always end with `\r\n`. The request parser
before would fault if a request ended only with `\n`, so this change
adds a quick check to handle both types of request. I'd have to check
the RFC to see which is actually correct.

While debugging, I added a pretty_print function that might be useful
for debugging moving forward.