Add a definition of gallocy::string
By adding a definition of gallocy::string we can start solving many of
the implicit conversion issues that have been so annoying to date. This
change adds a untested gallocy::string implementation (need to load test
for memory leaks still) and does not clean up the existing code base's
use of c_str() to side step the former problem.