Configure curl to use internal allocators.
It appears that the linking issue I solved earlier had something to do
with the curl errors that were screwing things up (we were compiling the
shared library against the system curl, and the rest of the application
against a locally compiled curl). After having had solved that problem,
the librypto free errors went away.