Refactor GallocyClient into GallocyMachine.
I believe the GallocyClient abstraction was misnamed and so I have
refactored the GallocyClient to be named GallocyMachine. This new
abstraction (per the docs) models the Raft state machine. This frees the
name GallocyClient for implementing an actual client -- e.g., an HTTP
client that knows the Gallocy protocol -- that can be used from the
GallocyMachine as well as the GallocyServer.

Naming is hard.